+++ src/Components/Withdraw/Domain/AccountRepository.php
@@ -4,7 +4,11 @@ namespace FelipeTorretto\ATM\Components\Withdraw\Domain;
 
 interface AccountRepository
 {
-    public function getBalance(int $id): float;
+    /**
+     * @param int $id
+     * @return int The balance in cents, e.g. 10000 for R$ 100,00
+     */
+    public function getBalance(int $id): int;
 
     public function updateBalance(int $id, float $balance): void;
 }
